Refactored Consumer Config and moved KafkaProducer to producer package
